Fixed plugin

+++ install-files/usr/share/n4d/python-plugins/ShutdownerClient.py
@@ -15,23 +15,41 @@ class ShutdownerClient:
 		self.core=n4dcore.Core.get_core()
 		self.cron_file="/etc/cron.d/lliurex-shutdowner"
 		self.desktop_cron_file="/etc/cron.d/lliurex-shutdowner-desktop"
-		self.override_folder="/etc/lliurex-shutdowner"
-		self.override_token=os.path.join(self.override_folder,"client-override.token")
+		self.override_shutdown_folder="/etc/lliurex-shutdowner"
+		self.override_shutdown_token=os.path.join(self.override_shutdown_folder,"client-override_shutdown.token")
 		self.adi_client="/usr/bin/natfree-client"
 		self.version_reference=["adi","desktop"]
 	
 	#def init
-	
+
 	def startup(self,options):
 		
-		if self._is_client_mode():
-			if os.path.exists(self.desktop_cron_file):
-				os.remove(self.desktop_cron_file)
-			t=threading.Thread(target=self._startup)
-			t.daemon=True
-			t.start()
+		t=threading.Thread(target=self._check_connection)
+		t.daemon=True
+		t.start()
 		
 	#def startup
+
+	def _check_connection(self):
+		
+		if self._is_client_mode():
+			remove_cron=True
+			max_retry=602
+			count=0
+			while True:
+				if count>=max_retry:
+					break
+				count+=1
+				time.sleep(1)
+				
+			if not self._check_connection_with_adi():
+				remove_cron=False
+			
+			if remove_cron:
+				if os.path.exists(self.desktop_cron_file):
+					os.remove(self.desktop_cron_file)
+			
+				self._startup()
 	
 	def _startup(self):
 		
@@ -42,7 +60,6 @@ class ShutdownerClient:
 		for x in range(0,tries):
 		
 			self.shutdowner_var=self.core.get_variable("SHUTDOWNER")["return"]
-			
 			if self.shutdowner_var != None:
 				self.shutdowner_trigger(self.shutdowner_var)
 				break
@@ -57,17 +74,18 @@ class ShutdownerClient:
 	
 	def shutdowner_trigger(self,value):
 		
-		override_shut=False
-		override_enabled=self._is_override_enabled()
+		override_shutdown=False
+		override_shutdown_enabled=self._is_override_shutdown_enabled()
 
 		if value!=None:
 			if not value["cron_values"]["server_shutdown"]:
-				if override_enabled:
-					override_shut=True
+				if override_shutdown_enabled:
+					override_shutdown=True
 
-			if not override_shut:
+			if not override_shutdown:
 				if value["cron_enabled"]:
 					if value["cron_content"]!=None:
+						value["cron_content"]=value["cron_content"].replace("&gt;&gt;",">>")
 						self._create_cron_file(value)
 				else:
 					self._delete_cron_file()
@@ -87,7 +105,7 @@ class ShutdownerClient:
 
 	def is_shutdown_override_enabled(self):
 
-		is_enabled=self._is_override_enabled()
+		is_enabled=self._is_override_shutdown_enabled()
 		
 		return n4d.responses.build_successful_call_response(is_enabled)
 
@@ -98,7 +116,7 @@ class ShutdownerClient:
 		value=self.core.get_variable("SHUTDOWNER")["return"]
 		ret=False
 		if not value["cron_values"]["server_shutdown"]:
-			self._create_override_token()
+			self._create_override_shutdown_token()
 			self._delete_cron_file()
 			ret=True
 		
@@ -108,29 +126,31 @@ class ShutdownerClient:
 
 	def disable_override_shutdown(self):
 
-		self._delete_override_token()
+		self._delete_override_shutdown_token()
 		value=self.core.get_variable("SHUTDOWNER")["return"]
 
 		if value!=None:
 			if value["cron_enabled"]:
 				if value["cron_content"]!=None:
+					value["cron_content"]=value["cron_content"].replace("&gt;&gt;",">>")
+
 					self._create_cron_file(value)
 
 		return n4d.responses.build_successful_call_response()
 
 	#def disable_override_shutdown
 
-	def _is_override_enabled(self):
+	def _is_override_shutdown_enabled(self):
 
-		if not os.path.exists(self.override_folder):
+		if not os.path.exists(self.override_shutdown_folder):
 			return False
 		else:
-			if os.path.exists(self.override_token):
+			if os.path.exists(self.override_shutdown_token):
 				return True
 			else:
 				return False
 
-	#def _is_override_enabled 
+	#def _is_override_shutdown_enabled 
 
 	def _create_cron_file(self,value):
 
@@ -147,23 +167,23 @@ class ShutdownerClient:
 
 	#def _delete_cron_file
 
-	def _create_override_token(self):
+	def _create_override_shutdown_token(self):
 
-		if not os.path.exists(self.override_folder):
-			os.mkdir(self.override_folder)
+		if not os.path.exists(self.override_shutdown_folder):
+			os.mkdir(self.override_shutdown_folder)
 
-		if not os.path.exists(self.override_token):
-			f=open(self.override_token,'w')
+		if not os.path.exists(self.override_shutdown_token):
+			f=open(self.override_shutdown_token,'w')
 			f.close()
 
-	#def _create_override_token
+	#def _create_override_shutdown_token
 
-	def _delete_override_token(self):
+	def _delete_override_shutdown_token(self):
 
-		if os.path.exists(self.override_token):
-			os.remove(self.override_token)
+		if os.path.exists(self.override_shutdown_token):
+			os.remove(self.override_shutdown_token)
 
-	#def _delete_override_token
+	#def _delete_override_shutdown_token
 	
 	def _is_client_mode(self):
 
@@ -190,8 +210,7 @@ class ShutdownerClient:
 							
 			if is_desktop:
 				if os.path.exists(self.adi_client):
-					if self._check_connection_with_adi():
-						is_client=True
+					is_client=True
 			
 			return is_client
